--Changed nanomachine radius to int. --Nanomachines now cubes of given radius with given center. --Set Nanomachine positions into the medium grid --Got rid of concept of destinations for moving molecules. --Molecules now check to see if there is a nanomachine of the appropriate type at their position in the grid to receive them.

+++ MolComSim/src/IntermediateNodeParam.java
@@ -9,7 +9,7 @@ public class IntermediateNodeParam extends NanoMachineParam {
 	private Position infoMolReleasePoint;
 	private Position ackMolReleasePoint;
 	
-	public IntermediateNodeParam(Position cntr, double r, Position infoPoint, Position ackPoint) {
+	public IntermediateNodeParam(Position cntr, int r, Position infoPoint, Position ackPoint) {
 		super(cntr, r, cntr);
 		infoMolReleasePoint = infoPoint;
 		ackMolReleasePoint = ackPoint;
